Replace tile layout with nstack
The nstack layout is a more general approach of the default tile layout. It has
a variable number of stack columns to the left and right of the master
area--controlled via the *nstackl*/*nstackr* variables. The width of the stack
area is evenly distributed onto the stacks.

Setting *nstackl* to 0 and *nstackr* to 1 makes the layout equivalent to the
default tile layout. Decreasing *nmaster* to 0 makes it a grid layout with an
adjustable number of columns.